Remove support for block weight
Block weights are now identical to their difficulties, so eliminate
the redundant weight concept. Keep block weight in the database block
struct, as removing it requires a migration. Also keep block weight in
the bootstrap serialization format, as removing it would be a
backwards-incompatible change, and should be done in a major release.